Address PR review comments on privileged-mode changes
- Split privileged-mode edge-case tests into a dedicated submodule
  to keep tests.rs under the 400-line limit
- Improve error messages in privileged_create helper for clearer
  test diagnostics
- Restore verification commands in 2-1-3 execplan that were
  collapsed into a single line by the markdown formatter
- Expand "BDD" acronym on first use in 2-2-2 execplan
- Fix Oxford comma and noun form in 2-2-2 execplan prose
