+++ docs/execplans/2-1-3-handle-socket-permission-errors.md
@@ -573,8 +573,15 @@ ______________________________________________________________________
 ## Verification
 
 ```bash
-set -o pipefail make check-fmt 2>&1 | tee /tmp/fmt.log make lint 2>&1 | tee
-/tmp/lint.log make test 2>&1 | tee /tmp/test.log
+make check-fmt
+```
+
+```bash
+make lint
+```
+
+```bash
+make test
 ```
 
 ## Testing notes
